webui/requirements.txt
streamlit>=1.37.0
Pillow>=10.0.0
